dont call enable versioning on join project

+++ services/web/app/coffee/Features/Templates/TemplatesController.coffee
@@ -23,6 +23,9 @@ module.exports =
 		request(zipUrl).pipe(writeStream)
 		writeStream.on 'close', ->
 			ProjectUploadManager.createProjectFromZipArchive req.session.user._id, req.session.templateData.templateName, dumpPath, (err, project)->
+				if err?
+					logger.err err:err, zipUrl:zipUrl, "problem building project from zip"
+					return res.send 500
 				setCompiler project._id, req.session.templateData.compiler, ->
 					fs.unlink dumpPath, ->
 					delete req.session.templateData
